[#51] Move Vault to core
