+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "2"
-members = [ "liberum_cli", "liberum_vault", "liberum_core", "liberum_gui"]
+members = [ "liberum_cli", "liberum_core", "liberum_gui"]
