+++ liberum_core/Cargo.toml
@@ -35,4 +35,9 @@ blake3 = "1.5.4"
 bs58 = "0.5.1"
 strum = "0.26"
 strum_macros = "0.26"
-uuid = {version="1.11", features=["serde"]}
+uuid = { version = "1.11", features = ["serde", "v4"] }
+pretty_assertions = "1.4.1"
+rand = "0.8.5"
+rusqlite = "0.32.1"
+tokio-rusqlite = "0.6.0"
+tokio-stream = "0.1.17"
