+++ liberum_core/src/main.rs
@@ -1,6 +1,7 @@
 pub mod connection;
 pub mod node;
 pub mod swarm_runner;
+pub mod vault;
 
 use anyhow::{anyhow, Result};
 use connection::listen;
