+++ liberum_core/src/vault/fragment.rs
