--- liberum_vault/.gitignore
@@ -1,1 +1,0 @@
-/target
