--- liberum_vault/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "liberum_vault"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-anyhow = "1.0.89"
-blake3 = "1.5.4"
-bs58 = "0.5.1"
-futures = "0.3.31"
-futures-util = "0.3.31"
-kameo = "0.12.2"
-liberum_core = { version = "0.1.0", path = "../liberum_core" }
-pretty_assertions = "1.4.1"
-rand = "0.8.5"
-rusqlite = { version = "0.32.1", features = ["bundled"] }
-tempdir = "0.3.7"
-tokio = { version = "1.40.0", features = ["full"] }
-tokio-rusqlite = "0.6.0"
-tokio-stream = "0.1.16"
-tokio-util = { version = "0.7.12", features = ["io"] }
-tracing = "0.1.40"
-uuid = { version = "1.11.0", features = ["v4"] }
